feat: exported the interface to a new file
