+++ migration-newip-to-ts/src/components/view/appView.ts
@@ -1,40 +1,6 @@
 import News from './news/news';
 import Sources from './sources/sources';
-
-interface IEverything {
-  status: 'ok' | 'error';
-  totalResults: number;
-  articles: IEverythingItem[];
-}
-
-interface IEverythingItem extends IEverything {
-  source: {
-    id: string;
-    name: string;
-  };
-  author: string;
-  title: string;
-  description: string;
-  url: string;
-  urlToImage: string;
-  publishedAt: string;
-  content: string;
-}
-
-interface ISources {
-  status: 'ok' | 'error';
-  sources: ISource[];
-}
-
-interface ISource extends ISources {
-  id: string;
-  name: string;
-  description: string;
-  url: string;
-  category: string;
-  language: string;
-  country: string;
-}
+import { IEverything, ISources } from '../types/interface';
 
 export class AppView {
   private news: News;
